refactor : content 분리
