+++ src/1_Page/ChampionshipEdit/constant/tab.ts
@@ -1,0 +1,6 @@
+export const CHAMPIONSHIP_EDIT_TAB = {
+  BASIC: "basic",
+  TEAMS: "teams",
+  AWARDS: "awards",
+  DATES: "dates",
+} as const;
