+++ src/1_Page/ChampionshipEdit/constant/type.d.ts
@@ -1,0 +1,1 @@
+type ChampionshipEditTab = "basic" | "teams" | "awards" | "dates";
