+++ src/1_Page/ChampionshipEdit/ui/AwardTab/type.d.ts
@@ -1,0 +1,3 @@
+type AwardTabProps = {
+  setActiveTab: (tab: ChampionshipEditTab) => void;
+};
