+++ src/1_Page/ChampionshipEdit/ui/TeamTab/model/useHandleTeamClick.ts
