+++ src/1_Page/ChampionshipEdit/ui/TeamTab/ui/TeamCard/index.tsx
