+++ src/1_Page/ChampionshipEdit/ui/TeamTab/ui/TeamCard/type.d.ts
